Correct permissions on git hooks and setup defaults
In order to properly run `prepare-commit-msg` in git repositories, this
commit makes all the hooks executable. This commit also sets up defaults
for the `web.browser` to be used with git based on the operating
system. This commit also updates the `workflow.zsh` file to xclip and
`bm_asana` for ticketing command.

[152380431446537](https://app.asana.com/0/152380431446537/152380431446537)